Add files via upload
This is version 9 - the final version of the scientific calculator application. Using NetBeans as an IDE, the Java Swing library for the GUIs, and several other packages this master file will run the entire application. It uses a master class-tester class style where the main class doing the heavy lifting "CoolCalc" is run via the main class titled the same name as the file "MyGUICalculatorOOP9".